more linear aggregation scheduling (all merges happen simultaniously)
